requirements.txt
smbus2==0.5.0
sqlite-web==0.6.4
RPi.GPIO==0.7.1